refactor(go): more like go project structure
