+++ lib/event.go
@@ -1,0 +1,5 @@
+package lib
+
+import "github.com/asaskevich/EventBus"
+
+var Event = EventBus.New()
